adding betty comments

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,10 +1,16 @@
 #include <limits.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include "lists.h"
+#include "listis.h"
+/**
+ * print_list - a function that prints all the elements of a list_t list.
+ * @h: the list.
+ *
+ * Return: the number of nodes.
+ */
 size_t print_list(const list_t *h)
 {
-	int size = 0;
+	size_t size = 0;
 
 	if (h->next == NULL)
 		return (0);
